+++ src/MVFC.Connectors.Developer/DaliboExplain/Modelos/Dtos/DaliboExplainUrlDto.cs
@@ -1,0 +1,5 @@
+﻿namespace MVFC.Connectors.Developer.DaliboExplain.Modelos.Dtos;
+
+public sealed record DaliboExplainUrlDto(
+        [property: JsonPropertyName("deleteKey")] string DeleteKey,
+        [property: JsonPropertyName("id")] string Id);