+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-﻿# Changelog
+# Changelog
 
 All notable changes to this project will be documented in this file.
 
@@ -6,6 +6,20 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ---
+
+## [3.0.3] - 2026-03-15
+
+### Added
+
+- `DaliboExplain` connector implementation in `MVFC.Connectors.Developer`
+
+### Removed
+
+- `MysqlExplain` connector implementation
+
+### Tests
+
+- Replaced `MysqlExplain` tests with `DaliboExplain` tests
 
 ## [3.0.2] -  2026-03-15
 
@@ -202,6 +216,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
+[3.0.3]: https://github.com/Marcus-V-Freitas/MVFC.Connectors/compare/v3.0.2...v3.0.3
+[3.0.2]: https://github.com/Marcus-V-Freitas/MVFC.Connectors/compare/v3.0.1...v3.0.2
 [3.0.1]: https://github.com/Marcus-V-Freitas/MVFC.Connectors/compare/v3.0.0...v3.0.1
 [3.0.0]: https://github.com/Marcus-V-Freitas/MVFC.Connectors/compare/v2.1.4...v3.0.0
 [2.1.4]: https://github.com/Marcus-V-Freitas/MVFC.Connectors/compare/v2.1.3...v2.1.4